requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
kafka-python==2.0.2
pydantic-settings==2.2.1
python-dotenv==1.0.1
aiokafka==0.9.0